Wire LEVERAGE, drop the drifted contract sizes, and make the docs checkable
`LEVERAGE=4` sat in docker-compose.yml, documented in AGENTS.md with a
default, and read by no code at all — so an operator lowering it watched the
book keep sizing at 4x. It multiplies target notional in
`execution.size_from_forecast` and divides margin, so the silence was
expensive. It is a `--leverage` flag now, on every research script through
`_common.py` and threaded from the orchestrator's `LEVERAGE` env default the
same way `EQUITY` already was. `Config.leverage` became a float, since
`--leverage 1.5` is a reasonable thing to ask for.

`test_every_environment_variable_compose_sets_is_read_by_something` is the
mechanism that stops the next one: every variable compose sets must appear
somewhere in backend/ or frontend/src, with an explicit exemption list for the
ones the runtime reads (TZ, the postgres entrypoint, vite). Verified by adding
a plausible `MAX_DRAWDOWN_HALT` and watching it fail.

DashboardPage carried a fallback contract-size table "until the specs load",
inherited from a cost module deleted in the rebuild. Three of its nine entries
had drifted against the real schedule — AVAX 10 against 5, LINK 50 against 10,
LTC 5 against 1 — and contract size multiplies straight into unrealised PnL, so
each misreported a position by that factor on every first render. There is no
fallback now; an instrument whose size is unknown shows the stored mark.
`PaperPositionsTable` took `?? 1` for the same value, which would have been a
100x error on XRP and 5000x on DOGE.

Docs: 41 stale claims corrected across AGENTS.md, CLAUDE.md, README.md and
docs/RESEARCH_PIPELINE.md. The ones that would have actively misled: AGENTS.md
listed `core/labels.py` as a live module three lines above the paragraph
describing what replaced it, documented six endpoints under a deleted router,
credited `costs.py` with the PnL and sizing layers it explicitly disclaims,
omitted `paper-engine` from both the service table and the "start everything"
command, and carried four environment variables no code reads while omitting
every one actually required to run. RESEARCH_PIPELINE.md now opens with a table
of the five places the build departed from the spec — labels, symbol identity,
the neural ensemble member, book-snapshot slippage, and the CPCV gate names —
rather than reading as current in sections that are not.

`test_docs.py` makes the checkable half of that a test: every path a doc names
must exist or be listed as deliberately-deleted with a reason. Verified by
adding a ghost module reference and watching it fail.

Tests are parallel by default (`-n auto` in pytest.ini): 5m50 across four cores
against 9m50 serially, and `-m "not slow"` is 33s. LightGBM is already pinned
to `n_jobs: 1`, so the workers do not oversubscribe. Two promotion evaluations
that run for a minute-plus were missing the `slow` mark, so the fast loop was
paying for them.

266 trader tests and 45 API tests pass; frontend typecheck, lint and build are
clean.

+++ backend/trader/core/config.py
@@ -102,7 +102,9 @@ class Config:
     # --- Risk / sizing ---
     max_positions: int = 5
     position_size: float = 0.15
-    leverage: int = 4
+    # Float, not int: `--leverage 1.5` is a reasonable thing for an operator to
+    # want, and `execution.py` already casts to float at both use sites.
+    leverage: float = 4.0
     vol_sizing_target: float = 0.025
     min_equity: float = 1000.0
